move conversation domain files from features/conversations into stores/conversation/
Stores own state and everything state needs; features keep UI and flows.
types, interfaces, schema, message-ordering, local-drafts, attachments (was
utils/attachments) and d1 (was adapters/d1) move with their tests. Boundary
TODO updated to name the two remaining stores->features imports.
