+++ src/stores/conversation/interfaces.ts
