+++ src/stores/conversation/schema.ts
