+++ src/stores/conversation/types.ts
